pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_tool_harness"
version = "0.1.0"
authors = [
  { name="zhu-weijie", email="byzhubaiyuan@gmail.com" },
]
description = "A library to easily build LLM agents with tool-using capabilities."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "anthropic~=0.52.0",
]

[project.optional-dependencies]
dev = [
    "build",
    "twine",
    "ruff",
    "pytest",
    "pytest-cov",
    "python-dotenv",
]
test = [
    "pytest",
    "pytest-cov",
]

[project.urls]
"Homepage" = "https://github.com/zhu-weijie/llm-tool-harness"
"Bug Tracker" = "https://github.com/zhu-weijie/llm-tool-harness/issues"